Resolved issues not resolved in feature/SQL
