+++ microGallery_Android/app/src/main/res/values/strings.xml
@@ -1,27 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
-    <string-array name="french_months">
-        <item>Janvier</item>
-        <item>Février</item>
-        <item>Mars</item>
-        <item>Avril</item>
-        <item>Mai</item>
-        <item>Juin</item>
-        <item>Juillet</item>
-        <item>Août</item>
-        <item>Septembre</item>
-        <item>Octobre</item>
-        <item>Novembre</item>
-        <item>Décembre</item>
+    <string-array name="months">
+        <item>January</item>
+        <item>February</item>
+        <item>March</item>
+        <item>April</item>
+        <item>May</item>
+        <item>June</item>
+        <item>July</item>
+        <item>August</item>
+        <item>September</item>
+        <item>October</item>
+        <item>November</item>
+        <item>December</item>
     </string-array>
     <string name="loading">Loading picture %s in webp</string>
     <string name="waitingForData">Downloading data</string>
     <string name="notLong">Should not take too long</string>
     <string name="calendar_title">%s %s</string>
     <string name="settings">Settings</string>
-    <string name="adress_of_server">IP adress of storage server</string>
-    <string name="ipv4">ip adress v4</string>
-    <string name="ipv6">ip adress v6</string>
+    <string name="adress_of_server">IP address of storage server</string>
+    <string name="ipv4">ip address v4</string>
+    <string name="ipv6">ip address v6</string>
     <string name="previsualisation_options">Preview option</string>
     <string name="preview_in_hd">Preview pictures in the highest definition</string>
     <string name="consumes_more">consumes more data</string>
