--- microGallery_Android/core-res/src/main/res/values/arrays.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-</resources>
